CJAL-19/12/2018- Avance Modulo Call center-pedidos

+++ 1.Layers/1.3Application/1.3.2MainModule/Application.MainModule/Servicios/AccesoADatos/TipoPerDataAccess.cs
@@ -22,7 +22,7 @@ namespace Application.MainModule.Servicios.AccesoADatos
             return uow.Repository<TipoPersona>().GetAll().ToList();
         }
 
-        public TipoPersona TiposPer(short persona)
+        public TipoPersona TiposPer(byte persona)
         {
             return uow.Repository<TipoPersona>().GetSingle(x=>x.IdTipoPersona.Equals(persona));
         }
